+++ changelog.md
@@ -6,6 +6,8 @@
   - Add a bunch new Pokemon
   - Add PvE against Champion Cynthia
   - Add form changing mechanic with DNA splicers or something
+    - Update getPokemonOrder
+    - Add a base Pokemon field or sm
   - Add Mythic Darkrai
     - Requires 30? tutorial stage completions
     - Gains {item} every 5 min, max 100 + level